+++ webapp.py
@@ -2,20 +2,25 @@ import utils
 import os
 from flask import Flask, render_template, flash, url_for, jsonify, request
 
-def webapp_thread(port_number):
+# encapsulates the functions for the web service so they can be run in a new thread
+def webapp_thread(port_number, debugMode=False):
     app = Flask(import_name="vsmp-plus", static_folder=os.path.join(utils.DIR_PATH, 'web', 'static'), template_folder=os.path.join(utils.DIR_PATH, 'web', 'templates'))
+
 
     @app.route('/', methods=['GET'])
     def index():
         return render_template('index.html', config=utils.get_configuration())
 
+
     @app.route('/setup', methods=['GET'])
     def setup_view():
         return render_template('setup.html', config=utils.get_configuration())
 
+
     @app.route('/api/configuration', methods=['GET'])
     def get_configuration():
         return jsonify(utils.get_configuration())
+
 
     @app.route('/api/configuration', methods=['POST'])
     def save_configuration():
@@ -53,6 +58,7 @@ def webapp_thread(port_number):
         result['message'] = 'Settings updated'
         return jsonify(result)
 
+
     @app.route('/api/control/<action>', methods=['POST'])
     def execute_action(action):
         result = {'success': True, 'message': ''}
@@ -62,6 +68,7 @@ def webapp_thread(port_number):
             config = utils.get_configuration()
             config['running'] = action == 'play'  # eval to True/False
             utils.write_json(utils.CONFIG_FILE, config)
+
             result['message'] = 'Action %s executed' % action
         else:
             result['success'] = False
@@ -74,4 +81,5 @@ def webapp_thread(port_number):
     def status():
         return jsonify(utils.read_json(utils.LAST_PLAYED_FILE))
 
-    app.run(debug=True, host='0.0.0.0', port=port_number, use_reloader=False)
+    # run the web app
+    app.run(debug=debugMode, host='0.0.0.0', port=port_number, use_reloader=False)